requirements.txt
# HTTP library for making requests
requests>=2.31.0

# SOCKS proxy support for Tor (required)
PySocks>=1.7.1